Move :with_fetched_tempfile to DuaController
